+++ src/Bootstrap.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Hirtz\Tenant;
 
+use Hirtz\Skeleton\Filters\PageCache;
 use Hirtz\Skeleton\Modules\Admin\Controllers\DashboardController;
 use Hirtz\Skeleton\Web\Application;
 use Hirtz\Skeleton\Web\UrlManager;
 use Hirtz\Tenant\Models\Tenant;
-use Hirtz\Tenant\Modules\Admin\Module;
 use Yii;
 use yii\base\BootstrapInterface;
 use yii\i18n\PhpMessageSource;
@@ -32,18 +32,43 @@ class Bootstrap implements BootstrapInterface
             Yii::$container->set(UrlManager::class, Web\UrlManager::class);
         }
 
-        $app->extendModule('admin', [
-            'modules' => [
-                'tenant' => [
-                    'class' => Module::class
-                ],
-            ],
+        if (!Yii::$container->has(PageCache::class)) {
+            Yii::$container->set(PageCache::class, Filters\PageCache::class);
+        }
+
+        $app->extendModule('tenant', [
+            'class' => Module::class,
         ]);
 
-        DashboardController::addRoles([
-            Tenant::AUTH_TENANT_UPDATE,
+        $app->extendComponent('sitemap', [
+            'variations' => function () {
+                $manager = Yii::$app->getUrlManager();
+                return $manager instanceof Web\UrlManager ? $manager->tenant?->id : null;
+            },
         ]);
+
+        if ($this->isAdminModuleEnabled()) {
+            $app->extendModule('admin', [
+                'modules' => [
+                    'tenant' => [
+                        'class' => Modules\Admin\Module::class,
+                    ],
+                ],
+            ]);
+
+            DashboardController::addRoles([
+                Tenant::AUTH_TENANT_UPDATE,
+            ]);
+        }
 
         $app->setMigrationNamespace('Hirtz\Tenant\Migrations');
+    }
+
+    /**
+     * @see Module::$enableAdminModule
+     */
+    protected function isAdminModuleEnabled(): bool
+    {
+        return (bool)(Yii::$app->getModules()['tenant']['enableAdminModule'] ?? true);
     }
 }